Update for Phoenix 1.2 in ecto model guides This includes issues 529, 530, and 531
Issues Handled:
Issue 529
https://github.com/phoenixframework/phoenix_guides/issues/529
This brought up the initial concerns about the various pieces that were
later broken into smaller issues after a suggestion by @jeregrine

Issue 530
https://github.com/phoenixframework/phoenix_guides/issues/530
These bits were specifically tied to the generators. They referenced the
now-unused `@required_fields` and `@optional_fields` rather than 1.2's
new style of declaration for this. All code references were changed by
using a project following the guide.

Issue 531
https://github.com/phoenixframework/phoenix_guides/issues/531
This removes the reference to the no-longer-generated line and also
corrects the statements around it to still be accurate. As there is
already plans mentioned for adding a section on scrubbing/securing
parameters properly, reference is most likely best left there in the
future.

The whole of the changes boil down to code references, and correction of
explanation of various functions used in the model pipeline. Controllers
and their portions as well as the installation-as-standalone pieces had
minimal modification.